EDA: Created 0.03, 0.04, modified 0.02, and created 1.01.
In 1.01, I created the interim data set that will be used for the modeling phase.
I also explored all distinct values in the data set to understand the data better (categorical variables, no date, location, etc.).

I also created 0.03 and 0.04 to initially explore the time data in the data set.
I will continue to explore the time data in the next commit.